reformatted to split user and post routers
