+++ server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const helmet = require("helmet");
-const dataRouter = require("./data/dataRouter");
+const postRouter = require("./data/postRouter");
+const userRouter = require("./data/userRouter");
 const server = express();
 
 function teamNamer(req, res, next) {
@@ -11,13 +12,12 @@ function teamNamer(req, res, next) {
 server.use(express.json());
 server.use(helmet());
 server.use(teamNamer);
-server.use("/api", dataRouter);
+server.use("/api/posts", postRouter);
+server.use("/api/users", userRouter);
 
 server.get("/", (req, res) => {
     res.send(`
-        <h2>Lambda Posts API</h2>
-
-        <p>Welcome ${req.team} to the Lambda Posts API</p>
+        <h2>Hi. You're at the wrong site! Visit /api/users or /api/posts to view the respective JSON objects.</h2>
         `);
 });
